rename subscription -> channel
