+++ src/rpc/mod.rs
@@ -5,6 +5,7 @@ mod auth_api;
 mod auth_layer;
 mod beacon_api;
 mod chain_api;
+mod channel;
 mod common_api;
 mod error;
 mod eth_api;
@@ -14,7 +15,6 @@ mod mpool_api;
 mod net_api;
 mod node_api;
 mod state_api;
-mod subscription;
 mod sync_api;
 mod wallet_api;
